Remove hard requirement on Unsafe
